Refactored LinearConverters into new module
