+++ examples/Ex01_simple/simple_example.py
@@ -10,7 +10,8 @@ import numpy as np
 import datetime
 
 from flixOpt import *
-from flixOpt.components import Boiler, Sink, Source, Storage, CHP
+from flixOpt.components import Sink, Source, Storage
+from flixOpt.linear_converters import Boiler, CHP
 
 ### some Solver-Inputs: ###
 displaySolverOutput = False # ausführlicher Solver-Output.
