+++ examples/Ex03_full_seg_agg/Model_and_solve.py
@@ -118,7 +118,8 @@ aTimeSeries = aTimeSeries.astype('datetime64')
 ##########################################################################
 
 from flixOpt import *
-from flixOpt.components import Boiler, CHP, Storage, Sink, Source
+from flixOpt.components import Storage, Sink, Source
+from flixOpt.linear_converters import Boiler, CHP
 
 print('#######################################################################')
 print('################### start of modeling #################################')
